Vaildate output and remove warnings

+++ bimodal.cc
@@ -45,14 +45,11 @@ int* bimodal(int m, char *tracefile, vector<int> table) {
     }
     InFile.close();
 
-//print_table(table);
- //       for (int i = 0; i < table.size(); i++)
-  //          cout << i << "\t" << table.at(i) << endl;
-
+    // Return results
     int *ret = (int *)malloc(sizeof(int) * (table.size() + 2));
     ret[0] = predictions;
     ret[1] = mispredictions;
-    for (int i = 0; i < table.size(); i++)
+    for (long unsigned int i = 0; i < table.size(); i++)
         ret[i+2] = table.at(i);
     return ret;
 }
